// webshop/src/models/Product.ts
export type Product = {
    id: number
    name: string
    price: number
    image: string
    category: string
    description: string
    active: boolean,
}

// export class Product {
//   constructor(
//     public id: number,
//     public name: string,
//     public price: number,
//     public image: string,
//     public category: string,
//     public description: string,
//     public active: boolean
//   ) {}
// }